// packages/components/button/src/type.d.ts
// 按钮类型类型
export type Type =
  | 'primary'
  | 'success'
  | 'warning'
  | 'danger'
  | 'info'
  | 'default'

// 按钮尺寸类型
export type Size = 'small' | 'default' | 'large'

// 按钮图标位置类型
export type IconPlacement = 'left' | 'right'

// 按钮原生类型类型
export type NativeType = 'button' | 'submit' | 'reset'